@@ -66,7 +66,7 @@
     "redux": "3.6.0",
     "redux-form": "^6.2.0",
     "redux-thunk": "2.1.0",
-    "rimraf": "^2.5.4",
+    "rimraf": "2.6.0",
     "sequelize": "3.24.7",
     "style-loader": "0.13.1",
     "superagent": "2.3.0",
@@ -78,16 +78,16 @@
     "xlsx": "^0.8.0"
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.1",
+    "babel-eslint": "7.1.1",
     "babel-plugin-istanbul": "^3.0.0",
     "chai": "^3.5.0",
     "cross-env": "^3.1.3",
-    "eslint": "^3.10.2",
-    "eslint-plugin-react": "6.7.1",
-    "eslint-config-opuscapita": "1.0.1",
+    "eslint": "3.15.0",
+    "eslint-plugin-react": "6.10.0",
+    "eslint-config-opuscapita": "1.0.4",
     "mocha": "^3.2.0",
     "mocha-junit-reporter": "^1.12.1",
-    "rimraf": "^2.5.4",
+    "npm-run-all": "4.0.1",
     "webpack-dev-middleware": "1.8.4",
     "webpack-hot-middleware": "2.13.1"
   }
