@@ -11,7 +11,7 @@ const campaignConstraints = {
       message: () => {return "^campaignEditor.campaignForm.error.isRequired"}
     }
   },
-  /*startsOn: {
+  /* startsOn: {
     presence: {
       message: () => {return "^campaignEditor.campaignForm.error.isRequired"}
     }
