@@ -4,26 +4,21 @@ import { createCampaign } from '../actions/campaigns/create';
 import { reduxForm } from 'redux-form';
 import { CREATE_CAMPAIGN_FORM } from '../constants/forms';
 import CampaignForm from '../components/CampaignEditor/CampaignForm.react';
-import { injectIntl, intlShape } from 'react-intl';
+import { intlShape } from 'react-intl';
 import { validateCampaign } from '../components/common/campaignValidator';
-import  campaignsType from '../../utils/workflowConstant.js';
+import campaignsType from '../../utils/workflowConstant.js';
 
 @connect(
   state => ({}),
-  (dispatch) => {
-    return {
-      handleCreateCampaign: function(router) {
-        dispatch(createCampaign(router));
-      }
-    }
-  }
+  (dispatch) => ({
+    handleCreateCampaign: (router) => dispatch(createCampaign(router))
+  })
 )
 export default class CampaignCreate extends Component {
-
   static propTypes = {
     intl: intlShape.isRequired,
     handleCreateCampaign: PropTypes.func.isRequired,
-    handleBack: PropTypes.func.isRequired,
+    onBack: PropTypes.func.isRequired,
     params: PropTypes.object
   };
 
@@ -33,7 +28,7 @@ export default class CampaignCreate extends Component {
   };
 
   render() {
-    const { intl, handleCreateCampaign } = this.props;
+    const { intl, handleCreateCampaign, onBack } = this.props;
     const { currentUserInfo: { username } } = this.context;
 
     return createElement(reduxForm({
@@ -43,7 +38,7 @@ export default class CampaignCreate extends Component {
       formLabel: intl.formatMessage({ id: 'campaignEditor.campaignForm.create.header' }),
       submitButtonLabel: intl.formatMessage({ id: 'campaignEditor.campaignForm.button.create' }),
       onSave: handleCreateCampaign.bind(null, this.context.router),
-      onCancel: ::this.props.handleBack,
+      onCancel: onBack,
       initialValues: { owner: username, status: 'new' },
       campaigntype: campaignsType.getWorkflowTypes()
     })(CampaignForm));
