@@ -1,19 +1,18 @@
-import React from 'react';
-import Thumbnail from '../components/common/Thumbnail.react';
-import { injectIntl, intlShape } from 'react-intl';
+import React, { Component } from 'react';
+import { intlShape } from 'react-intl';
 import _ from 'lodash';
 
-// loading from utils
+import Thumbnail from '../components/common/Thumbnail.react';
 import Template from '../../utils/template';
 
-export default class CampaignEmailTemplate extends React.Component {
+export default class CampaignEmailTemplate extends Component {
   constructor(props) {
     super(props);
     this.setDefaultState(props);
   }
 
   componentWillReceiveProps(nextProps) {
-    if (this.props.type != nextProps.type) {
+    if (this.props.type !== nextProps.type) {
       this.setDefaultState(nextProps);
     }
     this.props = nextProps;
@@ -23,41 +22,51 @@ export default class CampaignEmailTemplate extends React.Component {
     this.template = new Template();
     this.templates = this.template.get(props.type);
     this.defaultTemplate = this.template.getDefaultTemplate(props.type);
-    this.state = {selectedTemplate: this.defaultTemplate};
+    this.state = { selectedTemplate: this.defaultTemplate };
   }
 
   handleSelection = (templateId) => {
-    this.setState({selectedTemplate: templateId});
+    this.setState({ selectedTemplate: templateId });
+  }
+
+  handleBack = () => {
+    if (this.props.type === 'email') {
+      this.props.router.push(`/campaigns/edit/${this.props.campaignId}`);
+    } else {
+      this.props.router.push(`/campaigns/edit/${this.props.campaignId}/template/email`);
+    }
+  }
+
+  handleSave = () => {
+    if (this.props.type === 'email') {
+      this.props.router.push(`/campaigns/edit/${this.props.campaignId}/template/onboard`);
+    } else {
+      this.props.router.push(`/campaigns/edit/${this.props.campaignId}/contacts`);
+    }
   }
 
   renderThubmnails = () => {
     let thumbnails = [];
     for (let t in this.templates) {
       thumbnails.push(
-        <Thumbnail size={this.templates[t].size} onSelect={this.handleSelection.bind(this, this.templates[t].id)} isSelected={this.state.selectedTemplate == this.templates[t].id}  key={this.templates[t].id} src={this.templates[t].thumbnail} />
+        <Thumbnail
+          key={this.templates[t].id}
+          size={this.templates[t].size}
+          onSelect={this.handleSelection.bind(this, this.templates[t].id)}
+          isSelected={this.state.selectedTemplate === this.templates[t].id}
+          src={this.templates[t].thumbnail}
+        />
       );
     }
     return thumbnails;
   }
 
-  back = () => {
-    if (this.props.type == 'email')
-      this.props.router.push(`/campaigns/edit/${this.props.campaignId}`);
-    else
-      this.props.router.push(`/campaigns/edit/${this.props.campaignId}/template/email`);
-  }
-
-  save = () => {
-    if (this.props.type == 'email')
-      this.props.router.push(`/campaigns/edit/${this.props.campaignId}/template/onboard`);
-    else
-      this.props.router.push(`/campaigns/edit/${this.props.campaignId}/contacts`);
-  }
-
   render() {
+    const { type, intl: { formatMessage } } = this.props;
+
     return (
       <div className="form-horizontal">
-        <h1>{`Choose ${_.upperFirst(this.props.type)} Template`}</h1>
+        <h1>{`Choose ${_.upperFirst(type)} Template`}</h1>
         <div className="row">
           <div className="col-md-8">
             {this.renderThubmnails()}
@@ -65,14 +74,14 @@ export default class CampaignEmailTemplate extends React.Component {
         </div>
         <br />
         <div className="form-submit text-right">
-          <button className="btn btn-link" onClick={this.back}>
-            {this.props.intl.formatMessage({id: 'campaignEditor.steps.button.previous'})}
+          <button className="btn btn-link" onClick={this.handleBack}>
+            {formatMessage({ id: 'campaignEditor.steps.button.previous' })}
           </button>
           <button className="btn btn-primary">
-            {this.props.intl.formatMessage({id: 'campaignEditor.steps.button.createTemplate'})}
+            {formatMessage({ id: 'campaignEditor.steps.button.createTemplate' })}
           </button> &nbsp;
-          <button className="btn btn-primary" onClick={this.save}>
-            {this.props.intl.formatMessage({id: 'campaignEditor.steps.button.savenext'})}
+          <button className="btn btn-primary" onClick={this.handleSave}>
+            {formatMessage({ id: 'campaignEditor.steps.button.savenext' })}
           </button>
         </div>
       </div>
