@@ -1,13 +1,12 @@
-import React, { Component, PropTypes, createElement } from 'react';
+/* eslint-disable react/prop-types */
+
+import { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { Field, reduxForm } from 'redux-form';
 import { OnLoadCampaignPage, Onboarding } from '../actions/campaigns/onboard';
-import OnboardingCampaign from '../components/OnboardingCampaign.react';
-import { ONBOARDING_CAMPAIGN_FORM } from '../constants/forms';
-import { injectIntl, intlShape } from 'react-intl';
+import { injectIntl } from 'react-intl';
 
 @connect(
-  state => ({data: state.campaignList}),
+  state => ({ data: state.campaignList }),
   (dispatch) => {
     return {
       handleCampaignPageLoading: (campaignId, contactId, transition) => {
@@ -33,26 +32,26 @@ class CampaignPage extends Component {
   constructor(props) {
     super(props);
   }
-  
-  setCookie(cname,cvalue,exdays) {
-    var d = new Date();
-    d.setTime(d.getTime() + (exdays*24*60*60*1000));
-    var expires = "expires=" + d.toGMTString();
+
+  componentWillMount() {
+    const { campaignId, contactId } = this.props.params;
+
+    this.props.handleCampaignPageLoading(campaignId, contactId, this.props.location.query.transition);
+  }
+
+  setCookie(cname, cvalue, exdays) {
+    let d = new Date();
+    d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
+    let expires = "expires=" + d.toGMTString();
     document.cookie = cname + "=" + cvalue + ";" + expires + ";path=/";
     location.assign("/campaigns/ncc_onboard");
   }
-  componentWillMount() {
-    this.props.handleCampaignPageLoading(this.props.params.campaignId, this.props.params.contactId, this.props.location.query.transition);
-  }
-
-  
-  
 
   render() {
     if (this.props.data.onboardingCampaignContact !== undefined) {
-      var data ={
+      let data = {
         campaignId: this.props.params.campaignId,
-        contactId : this.props.params.contactId, 
+        contactId: this.props.params.contactId,
         contactCompany: this.props.data.onboardingCampaignContact.contact.companyName,
         companyId: 'ncc',
         companyName: 'NCC Svenska AB',
@@ -60,12 +59,12 @@ class CampaignPage extends Component {
         contactLastName: this.props.data.onboardingCampaignContact.contact.contactLastName,
         contactEmail: this.props.data.onboardingCampaignContact.contact.email
       }
-      var stringObj = JSON.stringify(data);
+      let stringObj = JSON.stringify(data);
       this.setCookie('CAMPAIGN_INFO', stringObj, 5)
     }
-    
-    const { intl, Onboarding } = this.props;
-    /*return createElement(reduxForm({
+
+    /* const { intl, Onboarding } = this.props;
+     return createElement(reduxForm({
       form: ONBOARDING_CAMPAIGN_FORM,
       fields: ['campaignId', 'contactId', 'transition'],
       initialValues: {
