@@ -1,4 +1,3 @@
-const _ = require('lodash');
 const async = require('async');
 const sendEmail = require('../../utils/emailIntegration');
 const workflowType = require('../../utils/workflowConstant');
@@ -7,7 +6,45 @@ let rule = new schedule.RecurrenceRule();
 rule.second = 0;
 
 module.exports = function(app, db) {
-  
+  // Get list of possible transitions.
+  const getTransitions = (campaignType, currentState) => {
+    return workflowType.getPossibleTransitions(campaignType, currentState);
+  }
+
+  // Update campaign contact's transition state.
+  const updateTransitionState = (campaignId, id, transitionState) => {
+    console.log('---campaignId---->', campaignId);
+    return db.CampaignContact.find({ where: { id: id } }).
+    then((contact) => {
+      if (contact && getTransitions('SupplierOnboarding', contact.dataValues.status).indexOf(transitionState) !== -1) {
+        return contact.updateAttributes({
+          status: transitionState
+        }).then((contact) => {
+          return contact;
+        });
+      } else {
+        return Promise.reject('Not possible to update transition.');
+      }
+    });
+  }
+
+  // To update Campaign Status.
+  const updateCampaignStatus = (campaignId, status) => {
+    return db.Campaign.find({ where: { campaignId: campaignId } }).
+    then((campaign) => {
+      return campaign.updateAttributes({
+        status: status
+      }).then((campaign) => {
+        return campaign;
+      }).catch((error) => {
+        return error;
+      })
+    }).
+    catch((error) => {
+      return error;
+    })
+  }
+
   /*
      API to get list of workflow.
   */
@@ -17,12 +54,12 @@ module.exports = function(app, db) {
   });
 
   app.post('/api/campaigns/start', (req, res) => {
-      db.sequelize.query("UPDATE Campaign SET status = 'new' WHERE campaignId = 'testNew'").spread( (results, metadata) => {
-        res.status(200).json(results);
-      })
+    const query = "UPDATE Campaign SET status = 'new' WHERE campaignId = 'testNew'";
+
+    db.sequelize.query(query).spread((results, metadata) => res.status(200).json(results));
   });
 
-  /*app.get('/api/getContacts', (req, res) => {
+  /* app.get('/api/getContacts', (req, res) => {
     db.CampaignContact.findAll({
       where: {status: 'read'},
       raw: true,
@@ -35,19 +72,18 @@ module.exports = function(app, db) {
     API to update the status of transition.
   */
   app.get('/api/transition/:campaignId/:contactId', (req, res) => {
-    db.Campaign.find({ where: {campaignId: req.params.campaignId}})
-    .then((campaign) => {
-      if(campaign){
-        updateTransitionState(campaign.type, req.params.contactId, req.query.transition)
-        .then((result) => {
-          res.status(200).json({campaign: campaign.dataValues, contact: result.dataValues});
+    db.Campaign.find({ where: { campaignId: req.params.campaignId } }).
+    then((campaign) => {
+      if (campaign) {
+        updateTransitionState(campaign.type, req.params.contactId, req.query.transition).
+        then((result) => {
+          res.status(200).json({ campaign: campaign.dataValues, contact: result.dataValues });
         }).catch((error) => {
-          res.status(500).json({message: 'Not able to update Transition status.'});
+          res.status(500).json({ message: 'Not able to update Transition status.' });
         })
-      }else{
-        res.status(500).json({message: 'There is no campaign of id'});
+      } else {
+        res.status(500).json({ message: 'There is no campaign of id' });
       }
-      
     })
   });
 
@@ -55,8 +91,8 @@ module.exports = function(app, db) {
    API to add onboard User's details.
   */
   app.post('/api/onboarding', (req, res) => {
-    updateTransitionState(req.body.campaignId, req.body.contactId, req.body.transition)
-    .then((result) => {
+    updateTransitionState(req.body.campaignId, req.body.contactId, req.body.transition).
+    then((result) => {
       res.status(200).json({});
     }).catch((error) => {
       res.status(500).json({});
@@ -67,58 +103,17 @@ module.exports = function(app, db) {
     API to quoued the list of contacts belogs to campaign.
   */
   app.put('/api/campaigns/start/:campaignId', (req, res) => {
-    updateCampaignStatus(req.params.campaignId, 'inprogress').then((data) =>{
-      db.sequelize.query("UPDATE CampaignContact SET status = 'queued' WHERE campaignId = '"+req.params.campaignId+"'").spread( (results, metadata) => {
-        res.status(200).json(data.dataValues);
-      })
+    updateCampaignStatus(req.params.campaignId, 'inprogress').then((data) => {
+      const query = `UPDATE CampaignContact SET status = 'queued' WHERE campaignId = '${req.params.campaignId}'`;
+      db.sequelize.query(query).spread((results, metadata) => res.status(200).json(data.dataValues));
     }).catch((error) => {
-      res.status(500).json({message: 'Not able to start campaign.'});
+      res.status(500).json({ message: 'Not able to start campaign.' });
     })
     ;
   });
-  
-  
-  //To update Campaign Status.
-  let updateCampaignStatus = (campaignId, status) => {
-    return db.Campaign.find({ where: {campaignId: campaignId}})
-    .then((campaign) => {
-      return campaign.updateAttributes({
-        status: status
-      }).then((campaign) => {
-          return campaign;
-      }).catch((error) => {
-        return error;
-      })     
-    })
-    .catch((error) => {
-      return error;
-    })
-  }
 
-  //Get list of possible transitions.
-  let getTransitions = (campaignType, currentState) => {
-    return workflowType.getPossibleTransitions(campaignType, currentState);
-  }
-  
-  //Update campaign contact's transition state. 
-  let updateTransitionState  = (campaignId, id, transitionState) => {
-    console.log('---campaignId---->', campaignId);
-    return db.CampaignContact.find({ where: {id: id} })
-    .then((contact) => {
-      if(contact && getTransitions('SupplierOnboarding', contact.dataValues.status).indexOf(transitionState)!== -1){
-        return contact.updateAttributes({
-          status: transitionState
-        }).then((contact) => {
-          return contact;
-        });
-      }else{
-        return Promise.reject('Not possible to update transition.');
-      }
-    });
-  }
-  
 
-  //API to send campaign emails.
+  // API to send campaign emails.
   let sendMails = () => {
     db.CampaignContact.findAll({
       where: {
@@ -126,24 +121,23 @@ module.exports = function(app, db) {
       },
       raw: true,
     }).then((contacts) => {
-      async.each(contacts, (contact, callback) => {  
+      async.each(contacts, (contact, callback) => {
         let sender = "opuscapita_noreply";
         let subject = "NCC Svenska AB asking you to connect eInvoicing";
-        sendEmail(sender, contact, subject, updateTransitionState, callback);                   
-      }, function(err){
-        if( err ) {
+        sendEmail(sender, contact, subject, updateTransitionState, callback);
+      }, function(err) {
+        if (err) {
           console.log('Not able to mail this --', err);
         } else {
           console.log('DONE');
         }
-      });               
+      });
     });
   }
-  
-  //Scheduler to send mails.
-  schedule.scheduleJob(rule, function(){
+
+  // Scheduler to send mails.
+  schedule.scheduleJob(rule, function() {
     sendMails();
   });
-
 };
 
