@@ -3,38 +3,38 @@ const _ = require('lodash');
 
 const workFlowsWithTransitions = [{
   'name': 'SupplierOnboarding',
-  'transitions':[{
+  'transitions': [{
 	  'name': 'queued',
-      'allowed': ['queued', 'sent', 'read', 'loaded', 'onboarded', 'bounced']
-	},{
+    'allowed': ['queued', 'sent', 'read', 'loaded', 'onboarded', 'bounced']
+  }, {
 	  'name': 'sent',
-      'allowed': ['sent', 'read', 'loaded', 'onboarded']
-	},{
+    'allowed': ['sent', 'read', 'loaded', 'onboarded']
+  }, {
 	  'name': 'read',
-      'allowed': ['read', 'loaded', 'onboarded']
-	},{
+    'allowed': ['read', 'loaded', 'onboarded']
+  }, {
 	  'name': 'loaded',
-      'allowed': ['loaded','onboarded']
-	},{
+    'allowed': ['loaded', 'onboarded']
+  }, {
 	  'name': 'onboarded',
-      'allowed': ['onboarded']
-	},{
+    'allowed': ['onboarded']
+  }, {
 	  'name': 'bounced',
-      'allowed': ['bounced', 'sent', 'read', 'loaded', 'onboarded']
-	}]
+    'allowed': ['bounced', 'sent', 'read', 'loaded', 'onboarded']
+  }]
 }]
 
 
 function getWorkflowTypes() {
-  let listOfWorkflows = _.map(workFlowsWithTransitions, function(workflow){
+  let listOfWorkflows = _.map(workFlowsWithTransitions, function(workflow) {
     return workflow.name;
   });
   return listOfWorkflows;
 }
 
-function getPossibleTransitions(workflowType, currentState){
- let workflowIndex = _.findKey(workFlowsWithTransitions, { 'name': workflowType});
- let transitionIndex = _.findKey(workFlowsWithTransitions[workflowIndex].transitions, { 'name': currentState});
+function getPossibleTransitions(workflowType, currentState) {
+  let workflowIndex = _.findKey(workFlowsWithTransitions, { 'name': workflowType });
+  let transitionIndex = _.findKey(workFlowsWithTransitions[workflowIndex].transitions, { 'name': currentState });
   return workFlowsWithTransitions[workflowIndex].transitions[transitionIndex].allowed;
 }
 
